requirements.txt
streamlit
yt-dlp
moviepy==1.0.3
google-generativeai
python-dotenv
selenium
webdriver-manager